Add ZPR viewer link option for  solr documents block; fixes #1704
