+++ app/assets/javascripts/spotlight/blocks/solr_documents_embed_block.js
@@ -1,8 +1,8 @@
-//= require spotlight/blocks/solr_documents_block
+//= require spotlight/blocks/solr_documents_base_block
 
 SirTrevor.Blocks.SolrDocumentsEmbed = (function(){
 
-  return SirTrevor.Blocks.SolrDocuments.extend({
+  return SirTrevor.Blocks.SolrDocumentsBase.extend({
     type: "solr_documents_embed",
 
     icon_name: "item_embed",
