+++ app/assets/javascripts/spotlight/blocks/solr_documents_features_block.js
@@ -1,8 +1,8 @@
-//= require spotlight/blocks/solr_documents_block
+//= require spotlight/blocks/solr_documents_base_block
 
 SirTrevor.Blocks.SolrDocumentsFeatures = (function(){
 
-  return SirTrevor.Blocks.SolrDocuments.extend({
+  return SirTrevor.Blocks.SolrDocumentsBase.extend({
     plustextable: false,
     type: "solr_documents_features",
